Move exporting of files into a pipeline stage
